Separated Ticket into tab components
