+++ client/src/components/TicketClosed.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function ClosedTickets() {
+  return (
+    <div>TicketClosed</div>
+  )
+}
