+++ client/src/components/TicketOpen.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function OpenTickets() {
+  return (
+    <div>TicketOpen</div>
+  )
+}
